zagent0/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "zagent0"
version = "0.1.0"
dependencies = [
  "aiqtoolkit[langchain]",
]
requires-python = ">=3.11,<3.13"
description = "Custom AIQ Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'aiq.components']
zagent0 = "zagent0.register"